pexpect: Make it possible to configure the maximum size of read
Using "ssh" or even "bash" as a command for a ExternalConsoleDriver
doesn't work because for some reason not all the received data is
returned by self._child.stdout.read(size).
Using bigger read fixes the problem.
An intermediate buffer could be used in ExternalConsoleDriver, but
increasing maxreadsize also fixes the issue and runs faster.

With ssh or bash as ExternalConsoleDriver, it possible to use the host
as any other DUT. The only remaining pitfall is to make sure that the
prompt can be parsed. To that effect, the safest is to define a very
simple PS1 (PS1='${debian_chroot:+($debian_chroot)}\u@\h:\w\$ ' is OK)

Signed-off-by: Jean-Jacques Hiblot <jjhiblot@traphandler.com>